+++ packages/table/src/table-header.js
@@ -16,47 +16,58 @@ export default {
             <colgroup
               name={ column.id }
               width={ column.realWidth || column.width }
-            />).concat(
-              <thead>
-                <tr>
-                  {
-                    this._l(this.columns, column =>
-                      <th
-                        on-mousemove={ ($event) => this.handleMouseMove($event, column) }
-                        on-mouseout={ this.handleMouseOut }
-                        on-mousedown={ ($event) => this.handleMouseDown($event, column) }
-                        on-click={ ($event) => this.handleHeaderClick($event, column) }
-                        class={ [column.id, column.direction, column.align] }>
-                        {
-                          [
-                            column.headerTemplate
-                              ? column.headerTemplate.call(this._renderProxy, h, column.label)
-                              : <div>{ column.label }</div>,
-                            column.sortable
-                              ? <div class="caret-wrapper">
-                                  <i class="sort-caret ascending"></i>
-                                  <i class="sort-caret descending"></i>
-                                </div>
-                              : ''
-                          ]
-                        }
-                      </th>
-                    ).concat(this.$parent.showVScrollBar && this.$parent.currentGutterWidth ? <th class="gutter"
-                              style={{ width: this.$parent.currentGutterWidth + 'px' }}></th> : '')
-                  }
-                </tr>
-              </thead>
-            )
+            />)
         }
+        {
+          !this.fixed && this.layout.gutterWidth
+            ? <colgroup name="gutter" width={ this.layout.scrollY ? this.layout.gutterWidth : '' }></colgroup>
+            : ''
+        }
+        <thead>
+          <tr>
+            {
+              this._l(this.columns, column =>
+                <th
+                  on-mousemove={ ($event) => this.handleMouseMove($event, column) }
+                  on-mouseout={ this.handleMouseOut }
+                  on-mousedown={ ($event) => this.handleMouseDown($event, column) }
+                  on-click={ ($event) => this.handleHeaderClick($event, column) }
+                  class={ [column.id, column.direction, column.align] }>
+                  {
+                    [
+                      column.headerTemplate
+                        ? column.headerTemplate.call(this._renderProxy, h, column.label)
+                        : <div>{ column.label }</div>,
+                      column.sortable
+                        ? <div class="caret-wrapper">
+                            <i class="sort-caret ascending"></i>
+                            <i class="sort-caret descending"></i>
+                          </div>
+                        : ''
+                    ]
+                  }
+                </th>
+              )
+            }
+            {
+              !this.fixed && this.layout.gutterWidth
+                ? <th class="gutter" style={{ width: this.layout.scrollY ? this.layout.gutterWidth + 'px' : '0' }}></th>
+                : ''
+            }
+          </tr>
+        </thead>
       </table>
     );
   },
 
   props: {
     columns: {},
-    fixed: Boolean,
-    allSelected: {
-      default: Boolean
+    fixed: String,
+    store: {
+      required: true
+    },
+    layout: {
+      required: true
     },
     border: Boolean
   },
@@ -66,9 +77,24 @@ export default {
     ElTag
   },
 
+  computed: {
+    isAllSelected() {
+      return this.store.states.isAllSelected;
+    },
+
+    columns() {
+      if (this.fixed === true || this.fixed === 'left') {
+        return this.store.states.fixedColumns;
+      } else if (this.fixed === 'right') {
+        return this.store.states.rightFixedColumns;
+      }
+      return this.store.states.columns;
+    }
+  },
+
   methods: {
-    toggleAllSelection($event) {
-      this.$parent.toggleAllSelection($event);
+    toggleAllSelection() {
+      this.store.commit('toggleAllSelection');
     },
 
     handleMouseDown(event, column) {
@@ -77,19 +103,19 @@ export default {
 
         this.$parent.resizeProxyVisible = true;
 
-        const gridEl = this.$parent.$el;
-        const gridLeft = gridEl.getBoundingClientRect().left;
+        const tableEl = this.$parent.$el;
+        const tableLeft = tableEl.getBoundingClientRect().left;
         const columnEl = this.$el.querySelector(`th.${column.id}`);
         const columnRect = columnEl.getBoundingClientRect();
-        const minLeft = columnRect.left - gridLeft + 30;
+        const minLeft = columnRect.left - tableLeft + 30;
 
         columnEl.classList.add('noclick');
 
         this.dragState = {
           startMouseLeft: event.clientX,
-          startLeft: columnRect.right - gridLeft,
-          startColumnLeft: columnRect.left - gridLeft,
-          gridLeft: gridLeft
+          startLeft: columnRect.right - tableLeft,
+          startColumnLeft: columnRect.left - tableLeft,
+          tableLeft
         };
 
         const resizeProxy = this.$parent.$refs.resizeProxy;
@@ -98,22 +124,20 @@ export default {
         document.onselectstart = function() { return false; };
         document.ondragstart = function() { return false; };
 
-        const mousemove = (event) => {
+        const handleMouseMove = (event) => {
           const deltaLeft = event.clientX - this.dragState.startMouseLeft;
           const proxyLeft = this.dragState.startLeft + deltaLeft;
 
           resizeProxy.style.left = Math.max(minLeft, proxyLeft) + 'px';
         };
 
-        const mouseup = () => {
+        const handleMouseUp = () => {
           if (this.dragging) {
             const finalLeft = parseInt(resizeProxy.style.left, 10);
             const columnWidth = finalLeft - this.dragState.startColumnLeft;
             column.width = column.realWidth = columnWidth;
 
-            this.$nextTick(() => {
-              this.$parent.$calcColumns();
-            });
+            this.store.scheduleLayout();
 
             document.body.style.cursor = '';
             this.dragging = false;
@@ -123,8 +147,8 @@ export default {
             this.$parent.resizeProxyVisible = false;
           }
 
-          document.removeEventListener('mousemove', mousemove);
-          document.removeEventListener('mouseup', mouseup);
+          document.removeEventListener('mousemove', handleMouseMove);
+          document.removeEventListener('mouseup', handleMouseUp);
           document.onselectstart = null;
           document.ondragstart = null;
 
@@ -133,8 +157,8 @@ export default {
           }, 0);
         };
 
-        document.addEventListener('mousemove', mousemove);
-        document.addEventListener('mouseup', mouseup);
+        document.addEventListener('mousemove', handleMouseMove);
+        document.addEventListener('mouseup', handleMouseUp);
       }
     },
 
@@ -146,13 +170,14 @@ export default {
       if (!this.dragging && this.border) {
         let rect = target.getBoundingClientRect();
 
+        var bodyStyle = document.body.style;
         if (rect.width > 12 && rect.right - event.pageX < 8) {
-          document.body.style.cursor = 'col-resize';
+          bodyStyle.cursor = 'col-resize';
           this.draggingColumn = column;
         } else if (!this.dragging) {
-          document.body.style.cursor = '';
+          bodyStyle.cursor = '';
           this.draggingColumn = null;
-          if (column.sortable) document.body.style.cursor = 'pointer';
+          if (column.sortable) bodyStyle.cursor = 'pointer';
         }
       }
     },
@@ -176,14 +201,14 @@ export default {
 
       if (!column.sortable) return;
 
-      const grid = this.$parent;
+      const sortCondition = this.store.states.sortCondition;
 
-      if (grid.sortingColumn !== column) {
-        if (grid.sortingColumn) {
-          grid.sortingColumn.direction = '';
+      if (sortCondition.column !== column) {
+        if (sortCondition.column) {
+          sortCondition.column.direction = '';
         }
-        grid.sortingColumn = column;
-        grid.sortingProperty = column.property;
+        sortCondition.column = column;
+        sortCondition.property = column.property;
       }
 
       if (!column.direction) {
@@ -192,25 +217,12 @@ export default {
         column.direction = 'descending';
       } else {
         column.direction = '';
-        grid.sortingColumn = null;
-        grid.sortingProperty = null;
+        sortCondition.column = null;
+        sortCondition.property = null;
       }
+      sortCondition.direction = column.direction === 'descending' ? -1 : 1;
 
-      grid.sortingDirection = column.direction === 'descending' ? -1 : 1;
-    },
-
-    $setVisibleFilter(property) {
-      if (this.visibleFilter) {
-        this.visibleFilter = null;
-      } else {
-        this.visibleFilter = property;
-      }
-    }
-  },
-
-  watch: {
-    visibleFilter(val) {
-      this.$parent.visibleFilter = val;
+      this.store.commit('changeSortCondition');
     }
   },
 
@@ -218,9 +230,7 @@ export default {
     return {
       draggingColumn: null,
       dragging: false,
-      dragState: {},
-      columnsMap: null,
-      visibleFilter: null
+      dragState: {}
     };
   }
 };